requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
pycryptodome
python-dotenv
